Week4-BigQuery/requirements.txt
# Function dependencies, for example:
# package>=version
pandas==1.4.2
google-cloud-storage==2.5.0
google-cloud-bigquery==3.3.5
requests==2.27.1

